SUS/SUS.MvcFramework.Tests/SusViewEngineTests.cs
using SUS.MVCFramework.ViewEngine;
using System;
using System.Collections.Generic;
using System.IO;
using Xunit;

namespace SUS.MvcFramework.Tests
{
    public class SusViewEngineTests
    {
        [Theory]
        [InlineData("CleanHtml")]
        [InlineData("Foreach")]
        [InlineData("IfElseFor")]
        [InlineData("ViewModel")]
        public void TestGetHtml(string fileName)
        {
            var viewModel = new TestViewModel()
            {
                DateOfBirth=new DateTime(2019, 6, 1),
                Name = "Doggo Argentino",
                Price = 12345.67M,
            };

            IViewEngine viewEngine = new SusViewEngine();
            string view = File.ReadAllText($"ViewTests/{fileName}.html");
            string actualResult= viewEngine.GetHtml(view, viewModel, null);
            string expectedResult = File.ReadAllText($"ViewTests/{fileName}.Result.html");
            Assert.Equal(expectedResult, actualResult); 
        }
        
        [Fact]
        public void TestTemplateViewModel()
        {
            IViewEngine viewEngine = new SusViewEngine();
            string actualResult=viewEngine.GetHtml(@"@foreach(var num in Model)
{
<span>@num</span>
}", new List<int> { 1, 2, 3 }, null);
            string expectedResult = @"<span>1</span>
<span>2</span>
<span>3</span>
";

            Assert.Equal(expectedResult, actualResult);
        }
    }
}
